DS-2570: remove original.subset and just use subset
I'm not sure why original.subset was used when missing = "Use partial
data" and a filter is provided (line 192 of the original code). The only
time original.subset differs from subset is when one of the weights is
zero or all values for a case are missing. When they do differ, a crash
occurs as we end up mixing vectors of different lengths. I have added
tests for these two cases, which used to fail but now pass with my
changes. Please also check the line "if (!(has.subset & (partial |
weighted))) result$cluster[subset] <- model$cluster # This will often do
nothing." as the condition seems strange to me.